fix(wgpu): embedding/cross_entropy_loss read index storage as raw bytes
Found while auditing cross_entropy_loss for autograd wiring. WGPU has
no genuine integer storage — index/target tensors are physically F32
bytes, and the embedding WGSL kernel confirms the correct way to read
them: it declares `indices: array<f32>` and does `u32(indices[i])`, a
real value conversion. But both embedding's backward
(`indices_capture.buffer.to_vec::<u32>()`) and cross_entropy_loss's
one-hot construction (`target.buffer.to_vec::<u32>()`) instead did a
raw bytemuck bit-reinterpret of those same bytes as u32 — which only
happens to produce the right answer for index/class 0.0, whose IEEE
bit pattern is 0x00000000. Any other value (e.g. 1.0f32 = bit pattern
1065353216) reads back as a huge garbage integer, which then fails
every `idx < vocab_size`/`class_idx < classes` bounds check and
silently drops that row's entire gradient/loss contribution instead
of erroring.

Existing tests never caught this: `embedding_backward_accumulates_gradients`
only used index 0.0 for both positions, and `test_cross_entropy_mean`
only asserted loose bounds (`loss > 0.0 && loss < 5.0`) despite using
target class 1 in one row. Verified this is real, not theoretical, by
temporarily reverting the fix and confirming both new regression tests
fail without it (embedding_backward_handles_nonzero_indices: wrong
gradient rows; cross_entropy_loss_matches_hand_computed_value_for_nonzero_target:
0.157 instead of the hand-computed 0.196) before restoring the fix and
confirming they pass. Full --features wgpu,std suite: 88/88 passing.

+++ crates/kindle-backends/src/wgpu/backend.rs
@@ -1891,7 +1891,13 @@ impl<T: DType, D: Device> ModuleOps<Self> for WgpuBackendImpl<T, D> {
             output_id: out_id,
             input_ids: vec![weight_id],
             backward: Box::new(move |grad_out: &WgpuStorage| {
-                let indices_data = indices_capture.buffer.to_vec::<u32>();
+                // Index storage is physically F32 bytes (the WGSL forward
+                // kernel declares `indices: array<f32>` and does a real
+                // `u32(indices[i])` value conversion) — `to_vec::<u32>()`
+                // would bit-reinterpret those bytes instead of converting
+                // the value, corrupting every index except 0.0 (whose bit
+                // pattern happens to equal integer 0).
+                let indices_data = indices_capture.buffer.to_vec::<f32>();
                 let grad_data = grad_out.buffer.to_vec::<f32>();
                 let mut weight_grad = vec![0.0f32; vocab_size * embed_dim];
 
@@ -2823,8 +2829,12 @@ impl<T: DType, D: Device> LossOps<Self> for WgpuBackendImpl<T, D> {
         // 1. log_probs = log_softmax(pred, 1)
         let log_probs = log_softmax::<T, K>(pred, 1)?;
 
-        // 2. Build one_hot constant WgpuStorage
-        let target_data = target.buffer.to_vec::<u32>();
+        // 2. Build one_hot constant WgpuStorage. Target storage is
+        // physically F32 bytes (this backend has no genuine integer
+        // storage), so read it as F32 and convert the value — a raw
+        // `to_vec::<u32>()` bit-reinterpret would corrupt every class index
+        // except 0.0 (whose bit pattern happens to equal integer 0).
+        let target_data = target.buffer.to_vec::<f32>();
         let mut one_hot_data = vec![0.0f32; batch * classes];
         for b_idx in 0..batch {
             let class_idx = target_data[b_idx] as usize;
